+++ src/tooling/piral-cli/src/injectors/pilet-injector.test.ts
@@ -7,6 +7,9 @@ const optionsMock = {
   api: '',
   app: '',
   active: true,
+  publicUrl: '/',
+  meta: '',
+  headers: {},
 };
 
 const configMock: any = {
@@ -85,6 +88,9 @@ describe('Piral-CLI piral injector', () => {
       pilets: [],
       api: 'http://someFakeApi:1234',
       app: '',
+      publicUrl: '/',
+      meta: '',
+      headers: {},
       active: true,
     };
 
@@ -112,6 +118,9 @@ describe('Piral-CLI piral injector', () => {
       pilets: [],
       api: 'http://someFakeApi:1234',
       app: '',
+      publicUrl: '/',
+      meta: '',
+      headers: {},
       active: true,
     };
     const core = new EventEmitter();