+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed issue with `piral-cli-esbuild` referencing files in CSS
 - Updated `kras` to `v0.14` and improved dev server config resolution
+- Added option to set `headers` in kras injector configuration
 
 ## 0.14.24 (May 13, 2022)